feat: initial support for connection via console_usb_serial_jtag
We can't use a uart callback when the usb-serial-jtag or usb-cdc console
is chosen as they don't support this, so we have to use base64|hex
encoding to upload/download files.

Unfortunatly, the current nodemcu firmware does not support sending long
chunks via usb-serial-jtag. See
https://github.com/nodemcu/nodemcu-firmware/issues/3657

Therefore, some features as 'file download' and 'send chunk of code from
editor window' do not work.

Also, if there are many files on spiffs, the extension hangs at the
stage of reading from ESP.