refactor: remove part of 'initialSettings'
The 'defaultValue' parameter from 'initialSettings' in
'WorkspaceConfiguration.get' is useful if the key doesn't exist at all
(e.g. if you access settings of an extension that is not present).
A value that should be returned when parameter could not be found, is
"undefined".
However, we are sure that 'package.json' always has the requested
parameter and its default value is defined.
So, we can remove "initial Settings", since the extension never accesses
them.