feat: add `Flash device`, `Erase flash` commands
This commit adds new commands 'Flash device' and 'Erase flash' to work
with flash using esptool.

To run the 'Flash ESP' and 'Erase flash' commands, you can use the icons
in the lower-left corner of the Status Bar. Or Ctrl+Shift+P -> NodeMCU:
Flash device or Erase flash

Before, you need to add the 'nodemcu-tools.flash' setting to
./vscode/settings.json file in the workspace. The setting is an object
with the properties:

 "esptool": command prompt for calling esptool.py
 "firmwareDir": the directory in workspace where the binary firmware
files are located. Default 'bin'
 "chipModel": the ESP model
 "baud": speed 115200 for ESP8266, 460800 for ESP32XX
 "target": This parameter is required only for esp8266. Different flash
sizes.

version of 'esptool' must be 4.9.0 or higher

Copy following files to 'firmwareDir':

- for esp32xx:
| file                | from nodemcu-firmware |
| ------------------- | --------------------- |
| nodemcu.bin         | build/nodemcu.bin |
| bootloader.bin      | build/bootloader/bootloader.bin |
| partition-table.bin | build/partition_table/partition-table.bin |
| flasher_args.json   | build/flasher_args.json |

- for esp8266:
  - 0x00000.bin
  - 0x10000.bin